Create shortcuts from manager
Delegate Intents callbacks to a single place.
